post-service/requirements.txt
# Core dependencies
fastapi==0.115.12
uvicorn==0.34.0
starlette==0.46.1
pydantic==2.11.1
pydantic_core==2.33.0

# Database dependencies
sqlalchemy
psycopg2

# Azure dependencies
azure-core==1.33.0
azure-storage-blob==12.25.1

# Security & encryption
bcrypt==4.3.0
cryptography==44.0.2
passlib==1.7.4
python-jose==3.4.0
ecdsa==0.19.1
rsa==4.9

# Environment variable management
python-dotenv==1.1.0

# Utility libraries
anyio==4.9.0
requests==2.32.3
h11==0.14.0
click==8.1.8
colorama==0.4.6
certifi==2025.1.31
idna==3.10
charset-normalizer==3.4.1
sniffio==1.3.1
six==1.17.0
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0

# Miscellaneous
annotated-types==0.7.0
cffi==1.17.1
pyasn1==0.4.8
pycparser==2.22
python-multipart==0.0.5